매장 폴더 분리
